requirements.txt
pyspark
pyyaml
apache-flink 
kafka-python
faker